# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Alpogo API Documentation",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/event/{slug}/ticket": {
            "get": {
                "tags": [
                    "event"
                ],
                "summary": "A\u00f1ade un ticket al evento",
                "description": "",
                "operationId": "get_ticket_from_event",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "event"
                ],
                "summary": "A\u00f1ade un ticket al evento",
                "description": "",
                "operationId": "add_ticket_to_event",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Ticket"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "408": {
                        "description": "Event not found"
                    },
                    "201": {
                        "description": "OK"
                    }
                }
            }
        },
        "/event/{slug}/ticket/{id}": {
            "put": {
                "tags": [
                    "event"
                ],
                "summary": "Edita un ticket de un evento",
                "description": "",
                "operationId": "update_ticket_of_event",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Ticket"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "408": {
                        "description": "Event not found"
                    },
                    "409": {
                        "description": "Ticket not found"
                    },
                    "201": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "event"
                ],
                "summary": "Edita parcialmente un ticket de un evento",
                "description": "",
                "operationId": "partial_update_ticket_of_event",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Ticket"
                        }
                    }
                ],
                "responses": {
                    "408": {
                        "description": "Event not found"
                    },
                    "409": {
                        "description": "Ticket not found"
                    },
                    "201": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/user/": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Detalle de usuario",
                "description": "Retorna el detalle de un usaurio, parametro de access token Authorization por header.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid id supplied",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "402": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Edita un usuario",
                "description": "",
                "operationId": "put_user",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de un usuario",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Invalid id supplied"
                    },
                    "402": {
                        "description": "User not found"
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Ticket": {
            "required": [
                "name",
                "stock",
                "amount",
                "producer_amount",
                "start_date",
                "end_date"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "El ticket de Jhon"
                },
                "stock": {
                    "type": "string",
                    "example": "21"
                },
                "amount": {
                    "type": "string",
                    "example": "123"
                },
                "producer_amount": {
                    "type": "string",
                    "example": "123"
                },
                "start_date": {
                    "type": "string",
                    "example": "2017/04/09"
                },
                "end_date": {
                    "type": "string",
                    "example": "2017/04/20"
                }
            },
            "type": "object",
            "xml": {
                "name": "Ticket"
            }
        },
        "User": {
            "required": [
                "first_name",
                "last_name",
                "email",
                "avatar",
                "passport",
                "birthday"
            ],
            "properties": {
                "first_name": {
                    "type": "string",
                    "example": "Jhon"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "email": {
                    "type": "string",
                    "example": "jhon@doe.com"
                },
                "avatar": {
                    "type": "string",
                    "example": "/jhon/doe/avatar.jpg"
                },
                "passport": {
                    "type": "string",
                    "example": "34335174"
                },
                "birthday": {
                    "type": "string",
                    "example": "1989/04/09"
                },
                "password": {
                    "type": "string",
                    "example": "jhondoe"
                },
                "password_confirmation": {
                    "type": "string",
                    "example": "jhondoe"
                }
            },
            "type": "object",
            "xml": {
                "name": "User"
            }
        }
    }
}